requirements.txt
slack_sdk==3.5.0
requests==2.25.1
fuzzywuzzy==0.18.0
python-dotenv==0.19.2
arrow==1.1.1
google-cloud-scheduler
google-cloud-functions
google-cloud-pubsub
